Refactor ClusterPersistRepositoryException
