+++ docs/document/content/user-manual/error-code/sql-error-code.en.md
@@ -84,6 +84,7 @@ SQL error codes provide by standard `SQL State`, `Vendor Code` and `Reason`, whi
 | HY000     | 17000       | Work ID assigned failed, which can not exceed 1024 |
 | HY000     | 17001       | Can not find \`%s\` file for datetime initialize |
 | HY000     | 17002       | File access failed, reason is: %s |
+| HY000     | 17010       | Cluster persist repository error, reason is: %s |
 
 ### Migration
